rename SavedBanner to SuccessBanner, fix number bug
