+++ Ghost_in_the_Shell/client_folder_react_app/ethisim/src/components/Banners/SuccessBanner.js
